+++ release-notes.md
@@ -4,6 +4,7 @@ Version 1.2.1
     state is retrieved from the ZoneManager and sent to the clients.
 * Added TimeSource.set/getOffset() to make it easier for clients to configure
     how far in history they'd like time to represent.  Defaults to -100 ms.
+* Expanded ZoneManager's javadoc.    
     
 
 Version 1.1.1